+++ py_css_styleguide/nomenclature.py
@@ -48,6 +48,5 @@ RESERVED_RULE_NAMES = (
 
 # Not validated, just here for mention
 RESERVED_PROPERTY_NAMES = (
-    #"flat",
     "structure",
 )